+++ src/CONST.js
@@ -1,0 +1,21 @@
+/*
+ * @Github: https://github.com/OBKoro1
+ * @Author: OBKoro1
+ * @Created_time: 2019-05-14 15:41:59
+ * @LastEditors: OBKoro1
+ * @LastEditTime: 2019-05-14 16:24:37
+ * @Description: 常量文件
+ */
+
+// 中间的注释符号，用以生成单行
+const middleAnnotation = {
+    javascript: ' * ',
+    python: '',
+    vb: "' ",
+    html: '* ',
+    // userLanguage: '' // 用户自定义的单行
+}
+
+module.exports = {
+    middleAnnotation,
+};